starter_cli/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "starter-cli"
version = "0.1.1"
description = "Operator CLI for the OpenAI Agents starter stack."
readme = "../README.md"
requires-python = ">=3.11"
license = { file = "../LICENSE" }
authors = [ { name = "OpenAI Agents Starter Team" } ]
keywords = ["cli", "agents", "starter"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "starter-contracts @ file:../starter_contracts",
  "httpx>=0.28.1,<0.29.0",
  "pydantic>=2.12.3,<2.13.0",
  "pyyaml>=6.0.2,<7.0.0",
  "python-dotenv>=1.0.0,<1.1.0",
  "rich>=13.9.0,<14.0.0",
  "boto3>=1.35.0,<2.0.0",
  "azure-identity>=1.17.0,<2.0.0",
  "azure-keyvault-secrets>=4.9.0,<5.0.0",
  "stripe>=10.11.0,<11.0.0",
  "textual>=0.64.0,<0.70.0",
  "openai-agents==0.6.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.3,<9.0.0",
  "pytest-asyncio>=0.21.1,<0.25.0",
  "fakeredis>=2.23.3,<3.0.0",
  "mypy>=1.11.0,<1.13.0",
  "pyright>=1.1.375,<1.2.0",
  "ruff>=0.7.0,<0.8.0",
  "types-PyYAML>=6.0.12.20240917,<7.0.0",
  # backend deps for shared test fixtures
  "fastapi>=0.115.12,<0.116.0",
  "pydantic>=2.12.3,<2.13.0",
  "pydantic-settings>=2.6.1,<2.7.0",
  "sqlalchemy>=2.0.35,<2.1.0",
  "asyncpg>=0.30.0,<0.31.0",
  "httpx>=0.28.1,<0.29.0",
  "uvicorn[standard]>=0.32.1,<0.33.0",
  "libpass[bcrypt]>=1.9.3,<2.0.0",
  "prometheus-client>=0.20.0,<0.21.0",
  "jsonschema>=4.22.0,<5.0.0",
  "jinja2>=3.1.3,<4.0.0",
  "zxcvbn>=4.4.0,<5.0.0",
  "email-validator>=2.1.0,<3.0.0",
  "geoip2>=4.7.0,<5.0.0",
  "resend>=2.13.1,<3.0.0",
  "aiosqlite>=0.20.0,<0.21.0",
  "greenlet>=3.0.3,<4.0.0",
]

[project.scripts]
starter-cli = "starter_cli.app:main"
aa-cli = "starter_cli.app:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.sources]
src = "src"

[tool.hatch.build.targets.wheel]
packages = ["starter_cli"]
exclude = ["**/__pycache__", "**/*.pyc"]
include = ["starter_cli/py.typed"]

[tool.hatch.build.targets.sdist]
include = [
  "src/**",
  "../starter_contracts/**",
  "../LICENSE",
  "../README.md",
]
exclude = ["**/__pycache__", "**/*.pyc"]

[tool.hatch.envs.default]
features = ["dev"]
python = "3.11"
path = ".venv"

[tool.hatch.envs.default.env-vars]
PYTHONPATH = "src:../starter_contracts/src"

[tool.hatch.envs.default.scripts]
lint = "ruff check src"
format = "ruff format src"
typecheck = "python -m pyright --project pyrightconfig.json && python -m mypy --config-file pyproject.toml src tests"
test = "pytest"
pyright = "python -m pyright --project pyrightconfig.json"

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP", "RUF"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_unused_ignores = true
strict_equality = true
namespace_packages = true
mypy_path = ["src", "../starter_contracts/src"]
disallow_any_generics = true
disallow_subclassing_any = true
ignore_missing_imports = true

[tool.pyright]
include = ["src"]
typeCheckingMode = "basic"
pythonVersion = "3.11"
extraPaths = ["src", "../starter_contracts/src"]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
pythonpath = ["src", "../starter_contracts/src", "../api-service/src"]
testpaths = ["tests"]
asyncio_mode = "auto"
